Option manager moves to controllers to disentangle circular deps
