+++ src/colmap/controllers/automatic_reconstruction.h
@@ -31,8 +31,8 @@
 
 #pragma once
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/scene/reconstruction_manager.h"
-#include "colmap/util/option_manager.h"
 #include "colmap/util/threading.h"
 
 #include <memory>
