+++ src/colmap/controllers/bundle_adjustment.h
@@ -31,8 +31,8 @@
 
 #pragma once
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/scene/reconstruction.h"
-#include "colmap/util/option_manager.h"
 #include "colmap/util/threading.h"
 
 namespace colmap {
