+++ src/colmap/controllers/option_manager.cc
@@ -29,7 +29,7 @@
 //
 // Author: Johannes L. Schoenberger (jsch-at-demuc-dot-de)
 
-#include "colmap/util/option_manager.h"
+#include "colmap/controllers/option_manager.h"
 
 #include "colmap/controllers/incremental_mapper.h"
 #include "colmap/estimators/bundle_adjustment.h"
