+++ src/colmap/controllers/option_manager.h
