+++ src/colmap/exe/feature.cc
@@ -32,13 +32,13 @@
 #include "colmap/exe/feature.h"
 
 #include "colmap/camera/models.h"
+#include "colmap/controllers/option_manager.h"
 #include "colmap/exe/gui.h"
 #include "colmap/feature/extraction.h"
 #include "colmap/feature/matching.h"
 #include "colmap/image/reader.h"
 #include "colmap/util/misc.h"
 #include "colmap/util/opengl_utils.h"
-#include "colmap/util/option_manager.h"
 
 namespace colmap {
 namespace {
