+++ src/colmap/exe/mvs.cc
@@ -31,12 +31,12 @@
 
 #include "colmap/exe/mvs.h"
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/mvs/fusion.h"
 #include "colmap/mvs/meshing.h"
 #include "colmap/mvs/patch_match.h"
 #include "colmap/scene/reconstruction.h"
 #include "colmap/util/misc.h"
-#include "colmap/util/option_manager.h"
 
 namespace colmap {
