+++ src/colmap/exe/vocab_tree.cc
@@ -31,6 +31,7 @@
 
 #include "colmap/exe/vocab_tree.h"
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/exe/gui.h"
 #include "colmap/feature/matching.h"
 #include "colmap/feature/sift.h"
@@ -39,7 +40,6 @@
 #include "colmap/scene/database.h"
 #include "colmap/util/misc.h"
 #include "colmap/util/opengl_utils.h"
-#include "colmap/util/option_manager.h"
 
 #include <numeric>
