+++ src/colmap/scene/reconstruction_manager.cc
@@ -32,7 +32,6 @@
 #include "colmap/scene/reconstruction_manager.h"
 
 #include "colmap/util/misc.h"
-#include "colmap/util/option_manager.h"
 
 namespace colmap {
 
@@ -66,8 +65,7 @@ size_t ReconstructionManager::Read(const std::string& path) {
   return idx;
 }
 
-void ReconstructionManager::Write(const std::string& path,
-                                  const OptionManager* options) const {
+void ReconstructionManager::Write(const std::string& path) const {
   std::vector<std::pair<size_t, size_t>> recon_sizes(reconstructions_.size());
   for (size_t i = 0; i < reconstructions_.size(); ++i) {
     recon_sizes[i] = std::make_pair(i, reconstructions_[i]->NumPoints3D());
@@ -83,9 +81,6 @@ void ReconstructionManager::Write(const std::string& path,
     const std::string reconstruction_path = JoinPaths(path, std::to_string(i));
     CreateDirIfNotExists(reconstruction_path);
     reconstructions_[recon_sizes[i].first]->Write(reconstruction_path);
-    if (options != nullptr) {
-      options->Write(JoinPaths(reconstruction_path, "project.ini"));
-    }
   }
 }
