+++ src/colmap/scene/reconstruction_manager.h
@@ -61,9 +61,7 @@ class ReconstructionManager {
   size_t Read(const std::string& path);
 
   // Write all managed reconstructions into sub-folders "0", "1", "2", ...
-  // If the option manager object is not null, the options are written
-  // to each respective reconstruction folder as well.
-  void Write(const std::string& path, const OptionManager* options) const;
+  void Write(const std::string& path) const;
 
  private:
   std::vector<std::shared_ptr<Reconstruction>> reconstructions_;
