+++ src/colmap/ui/bundle_adjustment_widget.h
@@ -31,10 +31,10 @@
 
 #pragma once
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/scene/reconstruction.h"
 #include "colmap/ui/options_widget.h"
 #include "colmap/ui/thread_control_widget.h"
-#include "colmap/util/option_manager.h"
 
 #include <QtCore>
 #include <QtWidgets>
