+++ src/colmap/ui/dense_reconstruction_widget.h
@@ -31,11 +31,11 @@
 
 #pragma once
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/mvs/fusion.h"
 #include "colmap/ui/image_viewer_widget.h"
 #include "colmap/ui/options_widget.h"
 #include "colmap/ui/thread_control_widget.h"
-#include "colmap/util/option_manager.h"
 
 #include <QtCore>
 #include <QtWidgets>
