+++ src/colmap/ui/feature_extraction_widget.h
@@ -31,8 +31,8 @@
 
 #pragma once
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/util/misc.h"
-#include "colmap/util/option_manager.h"
 
 #include <QtCore>
 #include <QtWidgets>
