+++ src/colmap/ui/image_viewer_widget.h
@@ -31,11 +31,11 @@
 
 #pragma once
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/geometry/projection.h"
 #include "colmap/scene/database.h"
 #include "colmap/scene/reconstruction.h"
 #include "colmap/ui/qt_utils.h"
-#include "colmap/util/option_manager.h"
 
 #include <QtCore>
 #include <QtWidgets>
