+++ src/colmap/ui/log_widget.h
@@ -31,7 +31,7 @@
 
 #pragma once
 
-#include "colmap/util/option_manager.h"
+#include "colmap/controllers/option_manager.h"
 
 #include <QtGui>
 #include <QtWidgets>
