+++ src/colmap/ui/match_matrix_widget.h
@@ -31,8 +31,8 @@
 
 #pragma once
 
+#include "colmap/controllers/option_manager.h"
 #include "colmap/ui/image_viewer_widget.h"
-#include "colmap/util/option_manager.h"
 
 namespace colmap {
